src/main/java/eu/volsch/stockmountain/extraction/conversion/AbstractStringConverter.java
/*
 * Copyright 2023 Volker Schmidt
 *
 * Redistribution and use in source and binary forms, with or without modification, are permitted
 * provided that the following conditions are met:
 *
 * 1. Redistributions of source code must retain the above copyright notice, this list of conditions
 *    and the following disclaimer.
 * 2. Redistributions in binary form must reproduce the above copyright notice, this list of
 *    conditions and the following disclaimer in the documentation and/or other materials provided
 *    with the distribution.
 * 3. Neither the name of the copyright holder nor the names of its contributors may be used to
 *    endorse or promote products derived from this software without specific prior written
 *    permission.
 *
 * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY EXPRESS OR
 * IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND
 * FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR
 * CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
 * DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
 * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
 * WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY
 * WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */

package eu.volsch.stockmountain.extraction.conversion;


import org.checkerframework.checker.nullness.qual.NonNull;
import org.checkerframework.checker.nullness.qual.Nullable;

/**
 * Abstract converter implementation that handles leading and trailing spaces and also converts to
 * <code>null</code> if resulting string is empty.
 *
 * @param <T> the target value type.
 */
public abstract class AbstractStringConverter<T> implements Converter<String, T> {

  @Override
  public final @NonNull Class<String> getSourceType() {
    return String.class;
  }

  @Override
  public final @Nullable T convert(@Nullable String source) throws ConversionException {
    if (source == null) {
      return null;
    }
    source = source.trim();
    if (source.isEmpty()) {
      return null;
    }
    return doConvert(source);
  }

  /**
   * Converts the specified non-null source value to the {@linkplain #getTargetType() target type}.
   *
   * @param source the source value that should be converted.
   * @return the converted source value.
   * @throws ConversionException thrown if the conversion cannot be performed.
   */
  protected abstract T doConvert(@NonNull String source) throws ConversionException;
}
